Make strings domain specific
